Added logic for swapping main components
