+++ src/main.ts
@@ -1,6 +1,7 @@
 /* eslint-disable */
 import Vue from 'vue'
 import App from "@/App.vue";
+import AppPackageBuilder from "@/AppPackageBuilder.vue";
 import VueRouter from 'vue-router'
 import router from "@/router"
 import store  from "@/store"
@@ -8,6 +9,7 @@ import vuetify from "./plugins/vuetify";
 import validation from "./plugins/validation";
 import { format, compareAsc } from 'date-fns';
 import InputMask from "inputmask";
+import createApp from 'vue'
 
 Vue.use(VueRouter);
 Vue.use(validation);
@@ -20,3 +22,10 @@ new Vue({
   vuetify,
   render: h => h(App),
 }).$mount('#app')
+
+// new Vue({
+//   store,
+//   router,
+//   vuetify,
+//   render: h => h(AppPackageBuilder),
+// }).$mount('#PackageBuilder')
