+++ src/store/index.ts
@@ -3,53 +3,10 @@ import Vuex from "vuex";
 
 Vue.use(Vuex);
 
-
 export default new Vuex.Store({
   plugins: [],
-  state: {
-    
-  },
-  /*
-    ███████████████████████████████████████████████████████████████████████████
-  
-    ███    ███ ██    ██ ████████  █████  ████████ ██  ██████  ███    ██ ███████
-    ████  ████ ██    ██    ██    ██   ██    ██    ██ ██    ██ ████   ██ ██
-    ██ ████ ██ ██    ██    ██    ███████    ██    ██ ██    ██ ██ ██  ██ ███████
-    ██  ██  ██ ██    ██    ██    ██   ██    ██    ██ ██    ██ ██  ██ ██      ██
-    ██      ██  ██████     ██    ██   ██    ██    ██  ██████  ██   ████ ███████
-  
-    ███████████████████████████████████████████████████████████████████████████
-    */
-  mutations: {
-  },
-  /*
-    ██████████████████████████████████████████████████████
-  
-     █████   ██████ ████████ ██  ██████  ███    ██ ███████
-    ██   ██ ██         ██    ██ ██    ██ ████   ██ ██
-    ███████ ██         ██    ██ ██    ██ ██ ██  ██ ███████
-    ██   ██ ██         ██    ██ ██    ██ ██  ██ ██      ██
-    ██   ██  ██████    ██    ██  ██████  ██   ████ ███████
-  
-    ██████████████████████████████████████████████████████
-    */
-  actions: {
-  },
-  /*
-    ██████████████████████████████████████████████████████████
-  
-     ██████  ███████ ████████ ████████ ███████ ██████  ███████
-    ██       ██         ██       ██    ██      ██   ██ ██
-    ██   ███ █████      ██       ██    █████   ██████  ███████
-    ██    ██ ██         ██       ██    ██      ██   ██      ██
-     ██████  ███████    ██       ██    ███████ ██   ██ ███████
-  
-    ██████████████████████████████████████████████████████████
-    */
-  getters: {
-     
-  },
+  state: {},
+  mutations: {},
+  actions: {},
+  getters: {},
 });
-  
-
-
