+++ types/Global.ts
